+++ chipcon/sram.h
@@ -36,8 +36,8 @@ void sram_write_byte(uint16_t addr, uint8_t data);
 //void sram_write_word(uint16_t addr, uint16_t data);
 uint8_t sram_read_byte(uint16_t addr);
 uint16_t sram_read_word(uint16_t addr);
-void sram_load_block(uint16_t addr,int size, uint8_t *b);
-void sram_store_block(uint16_t addr, int size, uint8_t *b);
+void sram_load_block(uint16_t addr,uint16_t size, uint8_t *b);
+void sram_store_block(uint16_t addr, uint16_t size, uint8_t *b);
 void sram_clear();
 
 #endif /* SRAM_H_ */
